ref(onboarding): Split JavaScript onboarding docs
