+++ static/app/gettingStartedDocs/javascript/javascript/profiling.tsx
@@ -1,0 +1,8 @@
+import {getJavascriptProfilingOnboarding} from 'sentry/utils/gettingStartedDocs/javascript';
+
+import {installSnippetBlock} from './utils';
+
+export const profiling = getJavascriptProfilingOnboarding({
+  installSnippetBlock,
+  docsLink: 'https://docs.sentry.io/platforms/javascript/profiling/browser-profiling/',
+});
