+++ static/app/gettingStartedDocs/javascript/javascript/utils.tsx
@@ -1,0 +1,619 @@
+import {Link} from '@sentry/scraps/link/link';
+
+import {buildSdkConfig} from 'sentry/components/onboarding/gettingStartedDoc/buildSdkConfig';
+import {
+  StepType,
+  type BasePlatformOptions,
+  type ContentBlock,
+  type DocsParams,
+  type OnboardingConfig,
+  type OnboardingStep,
+} from 'sentry/components/onboarding/gettingStartedDoc/types';
+import {
+  getAIRulesForCodeEditorStep,
+  getUploadSourceMapsStep,
+} from 'sentry/components/onboarding/gettingStartedDoc/utils';
+import {getFeedbackConfigOptions} from 'sentry/components/onboarding/gettingStartedDoc/utils/feedbackOnboarding';
+import {getReplayConfigOptions} from 'sentry/components/onboarding/gettingStartedDoc/utils/replayOnboarding';
+import {updateDynamicSdkLoaderOptions} from 'sentry/gettingStartedDocs/javascript/jsLoader/updateDynamicSdkLoaderOptions';
+import {t, tct} from 'sentry/locale';
+import {trackAnalytics} from 'sentry/utils/analytics';
+
+export enum InstallationMode {
+  AUTO = 'auto',
+  MANUAL = 'manual',
+}
+
+export const platformOptions = {
+  installationMode: {
+    label: t('Installation Mode'),
+    items: [
+      {
+        label: t('Loader Script'),
+        value: InstallationMode.AUTO,
+      },
+      {
+        label: t('Npm/Yarn'),
+        value: InstallationMode.MANUAL,
+      },
+    ],
+    defaultValue: InstallationMode.AUTO,
+  },
+} satisfies BasePlatformOptions;
+
+export type PlatformOptions = typeof platformOptions;
+export type Params = DocsParams<PlatformOptions>;
+
+export const isAutoInstall = (params: Params) =>
+  params.platformOptions.installationMode === InstallationMode.AUTO;
+
+const getIntegrations = (params: Params): string[] => {
+  const integrations = [];
+  if (params.isPerformanceSelected) {
+    integrations.push(`Sentry.browserTracingIntegration()`);
+  }
+
+  if (params.isProfilingSelected) {
+    integrations.push(`Sentry.browserProfilingIntegration()`);
+  }
+
+  if (params.isReplaySelected) {
+    integrations.push(
+      `Sentry.replayIntegration(${getReplayConfigOptions(params.replayOptions)})`
+    );
+  }
+
+  if (params.isFeedbackSelected) {
+    integrations.push(`
+      Sentry.feedbackIntegration({
+        colorScheme: "system",
+        ${getFeedbackConfigOptions(params.feedbackOptions)}
+      }),`);
+  }
+
+  return integrations;
+};
+
+const getDynamicParts = (params: Params): string[] => {
+  const dynamicParts: string[] = [];
+
+  if (params.isPerformanceSelected) {
+    dynamicParts.push(`
+      // Tracing
+      tracesSampleRate: 1.0, //  Capture 100% of the transactions
+      // Set 'tracePropagationTargets' to control for which URLs distributed tracing should be enabled
+      tracePropagationTargets: ["localhost", /^https:\\/\\/yourserver\\.io\\/api/]`);
+  }
+
+  if (params.isReplaySelected) {
+    dynamicParts.push(`
+      // Session Replay
+      replaysSessionSampleRate: 0.1, // This sets the sample rate at 10%. You may want to change it to 100% while in development and then sample at a lower rate in production.
+      replaysOnErrorSampleRate: 1.0 // If you're not already sampling the entire session, change the sample rate to 100% when sampling sessions where errors occur.`);
+  }
+
+  if (params.isProfilingSelected) {
+    dynamicParts.push(`
+        // Set profilesSampleRate to 1.0 to profile every transaction.
+        // Since profilesSampleRate is relative to tracesSampleRate,
+        // the final profiling rate can be computed as tracesSampleRate * profilesSampleRate
+        // For example, a tracesSampleRate of 0.5 and profilesSampleRate of 0.5 would
+        // results in 25% of transactions being profiled (0.5*0.5=0.25)
+        profilesSampleRate: 1.0`);
+  }
+
+  return dynamicParts;
+};
+
+export function getSdkSetupSnippet(params: Params) {
+  const config = buildSdkConfig({
+    params,
+    staticParts: [
+      `dsn: "${params.dsn.public}"`,
+      `// Setting this option to true will send default PII data to Sentry.
+      // For example, automatic IP address collection on events
+      sendDefaultPii: true`,
+    ],
+    getIntegrations,
+    getDynamicParts,
+  });
+
+  return `
+import * as Sentry from "@sentry/browser";
+
+Sentry.init({
+  ${config}
+});
+`;
+}
+
+export const installSnippetBlock: ContentBlock = {
+  type: 'code',
+  tabs: [
+    {
+      label: 'npm',
+      language: 'bash',
+      code: 'npm install --save @sentry/browser',
+    },
+    {
+      label: 'yarn',
+      language: 'bash',
+      code: 'yarn add @sentry/browser',
+    },
+    {
+      label: 'pnpm',
+      language: 'bash',
+      code: 'pnpm add @sentry/browser',
+    },
+  ],
+};
+
+export const verifySnippetBlock: ContentBlock[] = [
+  {
+    type: 'text',
+    text: t(
+      'This snippet includes an intentional error, so you can test that everything is working as soon as you set it up.'
+    ),
+  },
+  {
+    type: 'code',
+    tabs: [
+      {
+        label: 'JavaScript',
+        language: 'javascript',
+        code: `myUndefinedFunction();`,
+      },
+    ],
+  },
+];
+
+export function getAiRulesConfig(params: Params): OnboardingStep {
+  return getAIRulesForCodeEditorStep({
+    // ATTENTION: The rules defined here must match those in the documentation (see: https://github.com/getsentry/sentry-docs/blob/master/platform-includes/llm-rules-logs/javascript.mdx).
+    // If you make any changes, please update the docs accordingly.
+    rules: `
+These examples should be used as guidance when configuring Sentry functionality within a project.
+
+# Error / Exception Tracking
+
+Use \`Sentry.captureException(error)\` to capture an exception and log the error in Sentry.
+Use this in try catch blocks or areas where exceptions are expected
+
+# Tracing Examples
+
+Spans should be created for meaningful actions within an applications like button clicks, API calls, and function calls
+Ensure you are creating custom spans with meaningful names and operations
+Use the \`Sentry.startSpan\` function to create a span
+Child spans can exist within a parent span
+
+## Custom Span instrumentation in component actions
+
+\`\`\`javascript
+function handleTestButtonClick() {
+  // Create a transaction/span to measure performance
+  Sentry.startSpan(
+    {
+      op: "ui.click",
+      name: "Test Button Click",
+    },
+    (span) => {
+      const value = "some config";
+      const metric = "some metric";
+
+      // Metrics can be added to the span
+      span.setAttribute("config", value);
+      span.setAttribute("metric", metric);
+
+      doSomething();
+    },
+  );
+}
+\`\`\`
+
+## Custom span instrumentation in API calls
+
+\`\`\`javascript
+async function fetchUserData(userId) {
+  return Sentry.startSpan(
+    {
+      op: "http.client",
+      name: \`GET /api/users/\${userId}\`,
+    },
+    async () => {
+      const response = await fetch(\`/api/users/\${userId}\`);
+      const data = await response.json();
+      return data;
+    },
+  );
+}
+\`\`\`
+
+# Logs
+
+Where logs are used, ensure Sentry is imported using \`import * as Sentry from "@sentry/browser"\`
+Enable logging in Sentry using \`Sentry.init({ enableLogs: true })\`
+Reference the logger using \`const { logger } = Sentry\`
+Sentry offers a consoleLoggingIntegration that can be used to log specific console error types automatically without instrumenting the individual logger calls
+
+## Configuration
+
+### Baseline
+
+\`\`\`javascript
+import * as Sentry from "@sentry/browser";
+
+Sentry.init({
+  dsn: "${params.dsn.public}",
+
+  enableLogs: true,
+});
+\`\`\`
+
+### Logger Integration
+
+\`\`\`javascript
+Sentry.init({
+  dsn: "${params.dsn.public}",
+  integrations: [
+    // send console.log, console.warn, and console.error calls as logs to Sentry
+    Sentry.consoleLoggingIntegration({ levels: ["log", "warn", "error"] }),
+  ],
+});
+\`\`\`
+
+## Logger Examples
+
+\`logger.fmt\` is a template literal function that should be used to bring variables into the structured logs.
+
+\`\`\`javascript
+import * as Sentry from "@sentry/browser";
+
+const { logger } = Sentry;
+
+logger.trace("Starting database connection", { database: "users" });
+logger.debug(logger.fmt\`Cache miss for user: \${userId}\`);
+logger.info("Updated profile", { profileId: 345 });
+logger.warn("Rate limit reached for endpoint", {
+  endpoint: "/api/results/",
+  isEnterprise: false,
+});
+logger.error("Failed to process payment", {
+  orderId: "order_123",
+  amount: 99.99,
+});
+logger.fatal("Database connection pool exhausted", {
+  database: "users",
+  activeConnections: 100,
+});
+\`\`\`
+`,
+  });
+}
+
+export function getLoaderScriptCallbacks(params: Params) {
+  return {
+    onOptionalToggleClick: (showOptionalConfig: boolean) => {
+      if (showOptionalConfig) {
+        trackAnalytics('onboarding.js_loader_npm_docs_optional_shown', {
+          organization: params.organization,
+          platform: params.platformKey,
+          project_id: params.project.id,
+        });
+      }
+    },
+    onPageLoad: () => {
+      trackAnalytics('onboarding.setup_loader_docs_rendered', {
+        organization: params.organization,
+        platform: params.platformKey,
+        project_id: params.project.id,
+      });
+    },
+    onPlatformOptionsChange: () => {
+      trackAnalytics('onboarding.js_loader_npm_docs_shown', {
+        organization: params.organization,
+        platform: params.platformKey,
+        project_id: params.project.id,
+      });
+    },
+    onProductSelectionChange: ({previousProducts, products}: any) => {
+      updateDynamicSdkLoaderOptions({
+        orgSlug: params.organization.slug,
+        projectSlug: params.project.slug,
+        products,
+        previousProducts,
+        projectKey: params.projectKeyId,
+        api: params.api,
+      });
+    },
+    onProductSelectionLoad: (products: any) => {
+      updateDynamicSdkLoaderOptions({
+        orgSlug: params.organization.slug,
+        projectSlug: params.project.slug,
+        products,
+        projectKey: params.projectKeyId,
+        api: params.api,
+      });
+    },
+  };
+}
+
+const getVerifyConfig = () => [
+  {
+    type: StepType.VERIFY,
+    content: verifySnippetBlock,
+  },
+];
+
+export const loaderScriptOnboarding: OnboardingConfig<PlatformOptions> = {
+  introduction: () =>
+    tct("In this quick guide you'll use our [strong: Loader Script] to set up:", {
+      strong: <strong />,
+    }),
+  install: params => [
+    {
+      type: StepType.INSTALL,
+      content: [
+        {
+          type: 'text',
+          text: t('Add this script tag to the top of the page:'),
+        },
+        {
+          type: 'code',
+          tabs: [
+            {
+              label: 'HTML',
+              language: 'html',
+              code: `
+<script
+  src="${params.dsn.cdn}"
+  crossorigin="anonymous"
+></script>`,
+            },
+          ],
+        },
+        {
+          type: 'subheader',
+          text: t('Default Configuration'),
+        },
+        {
+          type: 'text',
+          text: t(
+            'The Loader Script settings are automatically updated based on the product selection above. Toggling products will dynamically configure the SDK defaults.'
+          ),
+        },
+        {
+          type: 'list',
+          items: [
+            tct(
+              'For Tracing, the SDK is initialized with [code:tracesSampleRate: 1], meaning all traces will be captured.',
+              {
+                code: <code />,
+              }
+            ),
+            tct(
+              'For Session Replay, the default rates are [code:replaysSessionSampleRate: 0.1] and [code:replaysOnErrorSampleRate: 1]. This captures 10% of regular sessions and 100% of sessions with an error.',
+              {
+                code: <code />,
+              }
+            ),
+          ],
+        },
+        {
+          type: 'text',
+          text: tct(
+            'You can review or change these settings in [link:Project Settings].',
+            {
+              link: (
+                <Link
+                  to={`/settings/${params.organization.slug}/projects/${params.project.slug}/loader-script/`}
+                />
+              ),
+            }
+          ),
+        },
+      ],
+    },
+  ],
+  configure: params => [
+    {
+      title: t('Configure SDK (Optional)'),
+      collapsible: true,
+      content: [
+        {
+          type: 'text',
+          text: t(
+            "Initialize Sentry as early as possible in your application's lifecycle."
+          ),
+        },
+        {
+          type: 'code',
+          tabs: [
+            {
+              label: 'HTML',
+              language: 'html',
+              code: `
+<script>
+  Sentry.onLoad(function() {
+    Sentry.init({${
+      params.isPerformanceSelected || params.isReplaySelected
+        ? ''
+        : `
+      // You can add any additional configuration here`
+    }${
+      params.isPerformanceSelected
+        ? `
+      // Tracing
+      tracesSampleRate: 1.0, // Capture 100% of the transactions`
+        : ''
+    }${
+      params.isReplaySelected
+        ? `
+      // Session Replay
+      replaysSessionSampleRate: 0.1, // This sets the sample rate at 10%. You may want to change it to 100% while in development and then sample at a lower rate in production.
+      replaysOnErrorSampleRate: 1.0, // If you're not already sampling the entire session, change the sample rate to 100% when sampling sessions where errors occur.`
+        : ''
+    }
+    });
+  });
+</script>`,
+            },
+          ],
+        },
+      ],
+      onOptionalToggleClick: showOptionalConfig => {
+        if (showOptionalConfig) {
+          trackAnalytics('onboarding.js_loader_npm_docs_optional_shown', {
+            organization: params.organization,
+            platform: params.platformKey,
+            project_id: params.project.id,
+          });
+        }
+      },
+    },
+    getAiRulesConfig(params),
+  ],
+  verify: getVerifyConfig,
+  nextSteps: () => [
+    {
+      id: 'source-maps',
+      name: t('Source Maps'),
+      description: t('Learn how to enable readable stack traces in your Sentry errors.'),
+      link: 'https://docs.sentry.io/platforms/javascript/sourcemaps/',
+    },
+  ],
+  onPageLoad: params => {
+    return () => {
+      trackAnalytics('onboarding.setup_loader_docs_rendered', {
+        organization: params.organization,
+        platform: params.platformKey,
+        project_id: params.project.id,
+      });
+    };
+  },
+  onPlatformOptionsChange: params => {
+    return () => {
+      trackAnalytics('onboarding.js_loader_npm_docs_shown', {
+        organization: params.organization,
+        platform: params.platformKey,
+        project_id: params.project.id,
+      });
+    };
+  },
+  onProductSelectionChange: params => {
+    return ({previousProducts, products}) => {
+      updateDynamicSdkLoaderOptions({
+        orgSlug: params.organization.slug,
+        projectSlug: params.project.slug,
+        products,
+        previousProducts,
+        projectKey: params.projectKeyId,
+        api: params.api,
+      });
+    };
+  },
+  onProductSelectionLoad: params => {
+    return products => {
+      updateDynamicSdkLoaderOptions({
+        orgSlug: params.organization.slug,
+        projectSlug: params.project.slug,
+        products,
+        projectKey: params.projectKeyId,
+        api: params.api,
+      });
+    };
+  },
+};
+
+export const packageManagerOnboarding: OnboardingConfig<PlatformOptions> = {
+  introduction: () =>
+    tct(
+      "In this quick guide you'll use [strong:npm], [strong:yarn], or [strong:pnpm] to set up:",
+      {
+        strong: <strong />,
+      }
+    ),
+  install: () => [
+    {
+      type: StepType.INSTALL,
+      content: [
+        {
+          type: 'text',
+          text: t(
+            "Sentry captures data by using an SDK within your application's runtime."
+          ),
+        },
+        installSnippetBlock,
+      ],
+    },
+  ],
+  configure: params => [
+    {
+      type: StepType.CONFIGURE,
+      content: [
+        {
+          type: 'text',
+          text: t(
+            "Initialize Sentry as early as possible in your application's lifecycle."
+          ),
+        },
+        {
+          type: 'code',
+          tabs: [
+            {
+              label: 'JavaScript',
+              language: 'javascript',
+              code: getSdkSetupSnippet(params),
+            },
+          ],
+        },
+      ],
+    },
+    getUploadSourceMapsStep({
+      guideLink: 'https://docs.sentry.io/platforms/javascript/sourcemaps/',
+      ...params,
+    }),
+    getAiRulesConfig(params),
+  ],
+  verify: getVerifyConfig,
+  nextSteps: () => [],
+  onPageLoad: params => {
+    return () => {
+      trackAnalytics('onboarding.js_loader_npm_docs_shown', {
+        organization: params.organization,
+        platform: params.platformKey,
+        project_id: params.project.id,
+      });
+    };
+  },
+  onProductSelectionChange: params => {
+    return ({products}) => {
+      updateDynamicSdkLoaderOptions({
+        orgSlug: params.organization.slug,
+        projectSlug: params.project.slug,
+        products,
+        projectKey: params.projectKeyId,
+        api: params.api,
+      });
+    };
+  },
+  onProductSelectionLoad: params => {
+    return products => {
+      updateDynamicSdkLoaderOptions({
+        orgSlug: params.organization.slug,
+        projectSlug: params.project.slug,
+        products,
+        projectKey: params.projectKeyId,
+        api: params.api,
+      });
+    };
+  },
+  onPlatformOptionsChange: params => {
+    return () => {
+      trackAnalytics('onboarding.setup_loader_docs_rendered', {
+        organization: params.organization,
+        platform: params.platformKey,
+        project_id: params.project.id,
+      });
+    };
+  },
+};
